Eliminate bootstrapper and seperate app from web
